refactor: simplify and clean up large-llm-response implementation
- Extract _TextFileTool base class to eliminate duplicate __init__ across read/search tools
- Inline case-insensitive matching in search tool to avoid full lowercased-lines copy
- Add cheap char-count pre-check in LargeResponseProcessor before UTF-8 encoding
- Compute content_size once instead of calling len(content_bytes) three times
- Fix notice message: report bytes not chars (len(content_bytes) is a byte count)
- Move json.loads and debug log inside if-tool guard in ToolExecutor.execute
- Remove unused size_threshold_override field from ProcessingContext
- Remove redundant comment on ResolvedConfig dataclass
