+++ src/quickapp/common/abstract/tool_call_result_processor.py
@@ -8,7 +8,6 @@ from quickapp.common.tool_call_result import ToolCallResult
 class ProcessingContext(BaseModel):
     tool_call_id: str | None
     tool_name: str
-    size_threshold_override: int | None = None
 
 
 class ToolCallResultProcessor(ABC):
