+++ src/quickapp/text_file_tooling/_base_text_file_tool.py
@@ -1,0 +1,28 @@
+from typing import Any
+
+from injector import AssistedBuilder
+
+from quickapp.common.perf_timer.perf_timer import PerformanceTimer
+from quickapp.common.staged_base_tool import StagedBaseTool
+from quickapp.config.tools.internal import InternalTool
+from quickapp.dial_core_services.dial_file_service import DialFileService
+from quickapp.text_file_tooling._stage_wrapper import _TextFileStageWrapper
+
+
+class _TextFileTool(StagedBaseTool):
+
+    def __init__(
+        self,
+        stage_wrapper_builder: AssistedBuilder[_TextFileStageWrapper],
+        tool_config: InternalTool,
+        perf_timer: PerformanceTimer,
+        dial_file_service: DialFileService,
+        **kwargs: Any,
+    ) -> None:
+        super().__init__(
+            stage_wrapper_builder=stage_wrapper_builder,  # type: ignore[arg-type]
+            tool_config=tool_config,
+            perf_timer=perf_timer,
+            **kwargs,
+        )
+        self._dial_file_service = dial_file_service
