+++ src/quickapp/text_file_tooling/_search_in_file_tool.py
@@ -1,34 +1,14 @@
 from typing import Any
 
-from injector import AssistedBuilder, inject
+from injector import inject
 
 from quickapp.common.base_stage_wrapper import BaseStageWrapper
-from quickapp.common.perf_timer.perf_timer import PerformanceTimer
-from quickapp.common.staged_base_tool import StagedBaseTool
 from quickapp.common.tool_call_result import ToolCallResult
-from quickapp.config.tools.internal import InternalTool
-from quickapp.dial_core_services.dial_file_service import DialFileService
-from quickapp.text_file_tooling._stage_wrapper import _TextFileStageWrapper
+from quickapp.text_file_tooling._base_text_file_tool import _TextFileTool
 
 
 @inject
-class _SearchInFileTool(StagedBaseTool):
-
-    def __init__(
-        self,
-        stage_wrapper_builder: AssistedBuilder[_TextFileStageWrapper],
-        tool_config: InternalTool,
-        perf_timer: PerformanceTimer,
-        dial_file_service: DialFileService,
-        **kwargs: Any,
-    ) -> None:
-        super().__init__(
-            stage_wrapper_builder=stage_wrapper_builder,  # type: ignore[arg-type]
-            tool_config=tool_config,
-            perf_timer=perf_timer,
-            **kwargs,
-        )
-        self._dial_file_service = dial_file_service
+class _SearchInFileTool(_TextFileTool):
 
     async def _run_in_stage_async(
         self,
@@ -44,10 +24,12 @@ class _SearchInFileTool(StagedBaseTool):
         content_bytes = await self._dial_file_service.download_file(file_url)
         lines = content_bytes.decode("utf-8").splitlines()
 
-        cmp_lines = [ln.lower() for ln in lines] if case_insensitive else lines
         cmp_pattern = pattern.lower() if case_insensitive else pattern
-
-        matching_indices = [i for i, ln in enumerate(cmp_lines) if cmp_pattern in ln]
+        matching_indices = [
+            i
+            for i, ln in enumerate(lines)
+            if cmp_pattern in (ln.lower() if case_insensitive else ln)
+        ]
 
         if not matching_indices:
             result = ToolCallResult(content="No matches found.", content_type="text/plain")
