+++ src/quickapp/tool_call_result_offload/_settings.py
@@ -6,8 +6,6 @@ from pydantic_settings import BaseSettings, SettingsConfigDict
 
 @dataclass(frozen=True)
 class ResolvedConfig:
-    # Default values come from env (ToolCallResultOffloadSettings).
-    # Individual fields can be overridden per-app via ToolCallResultOffloadAppConfig.
     enabled: bool
     size_threshold: int
     excluded_tools: frozenset[str]
